Added bf actions callback
