+++ build.gradle
@@ -162,7 +162,7 @@ dependencies {
     compile group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.12'
     compile group: 'uk.gov.hmcts.reform', name: 'document-management-client', version: '6.0.0'
 
-    compile group: 'uk.gov.hmcts.reform', name: 'ecm-common', version: '0.1.15'
+    compile group: 'uk.gov.hmcts.reform', name: 'ecm-common', version: '0.1.16'
 
     compile group: 'org.springframework.cloud', name: 'spring-cloud-starter-openfeign', version: '2.0.2.RELEASE'
     compile group: 'io.github.openfeign', name: 'feign-httpclient', version: '10.0.1'
